refactor(settings): move mock implementation from impl/ to mock/ package
