--- comp/core/settings/BUILD.bazel
@@ -1,10 +1,0 @@
-load("@rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "settings",
-    srcs = ["component_mock.go"],
-    importpath = "github.com/DataDog/datadog-agent/comp/core/settings",
-    tags = ["manual"],
-    visibility = ["//visibility:public"],
-    deps = ["//comp/core/settings/def"],
-)
