+++ comp/core/settings/impl/BUILD.bazel
@@ -1,11 +1,8 @@
-load("@rules_go//go:def.bzl", "go_library")
+load("@rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "impl",
-    srcs = [
-        "settings_mock.go",
-        "settingsimpl.go",
-    ],
+    srcs = ["settingsimpl.go"],
     importpath = "github.com/DataDog/datadog-agent/comp/core/settings/impl",
     visibility = ["//visibility:public"],
     deps = [
@@ -16,9 +13,23 @@ go_library(
         "//comp/def",
         "//pkg/config/model",
         "//pkg/util/scrubber",
-        "@com_github_gorilla_mux//:mux",
         "@com_github_json_iterator_go//:go",
         "@com_github_mohae_deepcopy//:deepcopy",
         "@in_yaml_go_yaml_v2//:yaml",
+    ],
+)
+
+go_test(
+    name = "impl_test",
+    srcs = ["settingsimpl_test.go"],
+    embed = [":impl"],
+    gotags = ["test"],
+    deps = [
+        "//comp/core/config",
+        "//comp/core/log/mock",
+        "//comp/core/settings/def",
+        "//pkg/config/model",
+        "@com_github_stretchr_testify//assert",
+        "@com_github_stretchr_testify//require",
     ],
 )
