+++ comp/remote-config/rcclient/impl/BUILD.bazel
@@ -40,6 +40,8 @@ go_test(
         "//comp/core/log/def",
         "//comp/core/log/mock",
         "//comp/core/settings/def",
+        "//comp/core/settings/fx",
+        "//comp/core/settings/mock",
         "//comp/core/sysprobeconfig/def",
         "//comp/remote-config/rcclient/def",
         "//pkg/config/mock",
@@ -49,8 +51,6 @@ go_test(
         "//pkg/remoteconfig/state",
         "//pkg/util/fxutil",
         "//pkg/util/log",
-        "@//comp/core/settings/fx",
-        "@//comp/core/settings/mock",
         "@com_github_stretchr_testify//assert",
         "@org_uber_go_fx//:fx",
     ],
